fixed cleanup method

+++ compute/cloud-client/src/test/java/compute/Util.java
@@ -187,7 +187,7 @@ public abstract class Util {
 
   // Delete reservations which starts with the given prefixToDelete and
   // has creation timestamp >24 hours.
-  public static void cleanUpExistingReservations(String projectId,
+  public static void cleanUpExistingReservations(String prefixToDelete, String projectId,
                                                  String zone)
       throws IOException, ExecutionException, InterruptedException, TimeoutException {
     try (ReservationsClient reservationsClient = ReservationsClient.create()) {
@@ -195,7 +195,9 @@ public abstract class Util {
         if (!reservation.hasCreationTimestamp()) {
           continue;
         }
-        DeleteReservation.deleteReservation(projectId, zone, reservation.getName());
+        if (reservation.getName().contains(prefixToDelete)) {
+          DeleteReservation.deleteReservation(projectId, zone, reservation.getName());
+        }
       }
     }
   }
